chore(lib): remove stray src/lib/lib folder
