+++ src/lib/constants.ts
@@ -1,0 +1,15 @@
+cat > lib/constants.ts <<'TS'
+export const COLORS = {
+  bg: '#ffffff',
+  text: '#111111',
+  accent: '#FFD54F',
+  black: '#000000',
+  line: '#eaeaea',
+  green: '#22c55e',
+  orange: '#f59e0b',
+  red: '#ef4444',
+  gray: '#e5e7eb',
+};
+
+export const NAV_HEIGHT = 84;
+TS
