+++ src/lib/thoughts.ts
